certmon module removed. Moved corresponding implementation into the prometheous module
